Added a button that prompts the user to specify the amount of squares per row they would like.
Added random colors to each square on the initial mouse over.
Each additional mouseover adds 10% more black to the square until it is
completly black after 15 mouseovers.

Added some more styling that centers the grid and the button to the
center of the screeen.

Further documented the code to assist in future readability.